glustercli: fix for local ip address issue in peerStatus
glustercli._getIpAddresses() throws an exception when
'ovirtmgmt' bridge is present. This is used by peerStatus() to
get the local ip address of the node and  peer status
is failing because of the exception.

Current approach is anyway not fool proof as the node can have
multiple network devices and we may not get the right ip address.
So adding 'localhost' as the address for the current host in
peerStatus() API.

Change-Id: Ib804b2109a2578f180b6bf54904a28f19bbdbb7a
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>
Reviewed-on: http://review.gluster.org/7688
Reviewed-by: Sahina Bose <sabose@redhat.com>