nagios-common: Georep cli session wise parsing corrected
Geo-rep status cli command does not uniquely identify
sessions per node line output. It prefixes the slave
node with slave volume name. And when the slave volume
has multiple nodes, the session status cannot be aggregated
based on this.
This patch strips out the slave node ip to uniquely identify
the session. This will not work when a volume has geo-rep
sessions created for multiple slave volumes - all with the
same name

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1101223
Change-Id: Ie6e12c1b3af4752c5ef14f0c7fe8db579013d19e
Signed-off-by: Sahina Bose <sabose@redhat.com>
Reviewed-on: http://review.gluster.org/7988
Reviewed-by: Shubhendu Tripathi <shtripat@redhat.com>
Reviewed-by: Kanagaraj M <kmayilsa@redhat.com>